+++ R/gini_index_fun.R
@@ -26,7 +26,7 @@
 #' @importFrom ineq Gini
 gini_index_fun <- function(x) {
 
-  x <- x[is.finite(x) & !is.na(x)]
+  x <- x[is.finite(x)]
   if (length(x) <= 1) return(0)
 
   ineq::Gini(x)