+++ tests/testthat.R
@@ -1,0 +1,3 @@
+library(testthat)
+library(softwareRisk)
+test_check("softwareRisk")